Lift the process-state isolation fixture to conftest, so main() tests cannot leak handlers
The four main() tests in tests/test_refetch_panos.py left four RotatingFileHandlers on the root logger -
measured 0 before, 4 after running the module alone - each pointed into a tmp_path pytest was about to
delete, with the root level left at DEBUG and the SIGTERM handler replaced. tests/test_download_runner.py
had the autouse fixture for exactly this, module-local, since #52.1; the next module to grow an in-process
main() test rediscovered the problem rather than the fixture.

It is now suite-wide in conftest.py, unchanged in what it snapshots and restores. Re-measured: 0 before,
0 after.
